refacored invoice generator

+++ InvoiceManagementSystem/Controllers/Product.cs
@@ -50,7 +50,7 @@ namespace InvoiceManagementSystem.Controllers
                                 float ProductTax = float.Parse(Console.ReadLine());
                                 Console.WriteLine("Enter the Product discount");
                                 double discount = Convert.ToDouble(Console.ReadLine());
-                                double ProductDiscount = discount / 100;
+                                float ProductDiscount =(float) discount / 100;
                                 Console.WriteLine("Enter the Category ID");
                                 string productCategoryiD = Console.ReadLine();
