+++ InvoiceManagementSystem/Services/Product.cs
@@ -13,7 +13,7 @@ namespace InvoiceManagementSystem.Services
             _product = ProductObject;
         }
        
-        public void AddProductsService(string productName, string productDescription, int productPrice, double productDiscount, float productTax, string categoryID)
+        public void AddProductsService(string productName, string productDescription, int productPrice, float productDiscount, float productTax, string categoryID)
         {
             _product.AddProducts(productName, productDescription, productPrice, productDiscount, productTax, categoryID);
         }
@@ -21,7 +21,7 @@ namespace InvoiceManagementSystem.Services
         {
             _product.DeleteProducts(productId);
         }
-        public void EditProductDetailsService(Models.Product Product,string productName,string productDescription,int productPrice,double productDiscount,float productTax   )
+        public void EditProductDetailsService(Models.Product Product,string productName,string productDescription,int productPrice,float productDiscount,float productTax   )
         {
             _product.EditProductDetails(Product,productName, productDescription, productPrice, productDiscount, productTax);
         }
